+++ oec/db_sitc/models.py
@@ -6,7 +6,7 @@ class Yo(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yo'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     origin_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
     import_val = db.Column(db.Numeric(16,2))
@@ -18,7 +18,7 @@ class Yd(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yd'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     dest_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
     import_val = db.Column(db.Numeric(16,2))
@@ -30,12 +30,12 @@ class Yp(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yp'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     sitc_id = db.Column(db.String(8), db.ForeignKey(Sitc.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
     import_val = db.Column(db.Numeric(16,2))
     pci = db.Column(db.Float())
-    pci_rank = db.Column(db.Integer(11))
+    pci_rank = db.Column(db.Integer)
     
     def __repr__(self):
         return '<Yp %d.%s>' % (self.year, self.sitc_id)
@@ -44,7 +44,7 @@ class Yop(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yop'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     origin_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     sitc_id = db.Column(db.String(8), db.ForeignKey(Sitc.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
@@ -60,7 +60,7 @@ class Yod(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yod'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     origin_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     dest_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
@@ -73,7 +73,7 @@ class Ydp(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_ydp'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     dest_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     sitc_id = db.Column(db.String(8), db.ForeignKey(Sitc.id), primary_key=True)
     export_val = db.Column(db.Numeric(16,2))
@@ -86,7 +86,7 @@ class Yodp(db.Model, AutoSerialize):
     
     __tablename__ = 'sitc_yodp'
     
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer, primary_key=True)
     origin_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     dest_id = db.Column(db.String(5), db.ForeignKey(Country.id), primary_key=True)
     sitc_id = db.Column(db.String(8), db.ForeignKey(Sitc.id), primary_key=True)
